+++ CHANGELOG.md
@@ -12,6 +12,7 @@ customers cannot upgrade their bootloader, its changes are recorded separately.
 - Remove option to restore from 18 recovery words
 - simulator: enable Test Merchant for payment requests
 - simulator: simulate a Nova device
+- Add API call to fetch multiple xpubs at once
 
 ### 9.23.1
 - EVM: add HyperEVM (HYPE) and SONIC (S) to known networks